feat: recipe-screen.tsxを分割
- lib/recipe-parser.ts に食材解析ロジックを切り出し
- recipe-screen.tsx をコンテナに絞り込み（view切り替え・recipeList・customIngredients管理のみ）
- recipe-list-view.tsx / recipe-detail-view.tsx / recipe-edit-view.tsx にビューを分離
- recipe-detail-view の削除ボタンを bottom-16 に修正（BottomNav重複対応）

Closes #61
